.claude/skills/openai-chatkit-integration/assets/backend/pyproject.toml
[project]
name = "chatkit-backend"
version = "0.1.0"
description = "Production-ready ChatKit backend with OpenAI Agents SDK and alternative LLM providers"
requires-python = ">=3.13"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # ChatKit SDK (Official Python SDK)
    "openai-chatkit>=1.4.2",

    # OpenAI Agents SDK
    "openai-agents>=0.6.5",
    "openai>=1.57.0",  # AsyncOpenAI client

    # Data Validation & Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Environment
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",

    # Optional: Database (uncomment for persistent storage)
    # "sqlmodel>=0.0.22",
    # "asyncpg>=0.30.0",  # PostgreSQL async driver
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",  # For testing FastAPI
]
